FIX: Correct multiply typos
