+++ quantecon/game_theory/__init__.py
@@ -10,10 +10,5 @@ from .support_enumeration import support_enumeration, support_enumeration_gen
 from .lemke_howson import lemke_howson
 from .mclennan_tourky import mclennan_tourky
 from .vertex_enumeration import vertex_enumeration, vertex_enumeration_gen
-from .utilities import NashResult, RGUtil
-from .repeated_game import (
-	 flow_u_1, flow_u_2, flow_u, best_dev_i, best_dev_1, best_dev_2, 
-	 best_dev_payoff_i, best_dev_payoff_1, best_dev_payoff_2, initialize_hpl,
-	 worst_value_i, worst_value_1, worst_value_2, worst_values, RepeatedGame,
-	 outerapproximation
-)
+from .utilities import NashResult
+from .repeated_game import RepeatedGame, outerapproximation
