+++ quantecon/game_theory/tests/test_repeated_game.py
@@ -6,12 +6,14 @@ Tests for repeated_game.py
 """
 
 import numpy as np
+from math import pi, cos, sin
+from numpy.testing import assert_array_equal
 from nose.tools import ok_
-from quantecon.game_theory import NormalFormGame, Player, flow_u_1, flow_u_2, \
-                                  flow_u, best_dev_i, worst_value_i, \
-                                  initialize_hpl, outerapproximation, \
-                                  RepeatedGame, RGUtil
-
+from quantecon.game_theory import NormalFormGame, Player
+from ..repeated_game import (
+     flow_u_1, flow_u_2, flow_u, best_dev_i, unitcircle, worst_value_i, 
+     initialize_hpl, outerapproximation, RepeatedGame,
+)
 
 class TestRepeatedGame:
     def setUp(self):
@@ -25,6 +27,22 @@ class TestRepeatedGame:
         self.rpd = RepeatedGame(nfg, 0.75)
         self.C, self.H, self.Z = initialize_hpl(4, [0.0, 0.0], 1.0)
 
+    def test_unitcircle(self):
+        incr = 2*pi/5
+        pts = np.array([[cos(0. * incr), sin(0. * incr)],
+                        [cos(1. * incr), sin(1. * incr)],
+                        [cos(2. * incr), sin(2. * incr)],
+                        [cos(3. * incr), sin(3. * incr)],
+                        [cos(4. * incr), sin(4. * incr)]])
+
+        npts = 5
+        test_obj = unitcircle(npts)
+        assert_array_equal(test_obj, pts)
+
+        H = unitcircle(4)
+        points = np.array([[1, 0], [0, 1], [-1, 0], [0, -1]])
+        ok_(max(abs(H - points).flatten()) < 1e-12)
+
     def test_flow_utilities(self):
         ok_(abs(flow_u_1(self.rpd, 0, 0) - 9.0) < 1e-14)
         ok_(abs(flow_u_2(self.rpd, 1, 1) - 3.0) < 1e-14)
@@ -32,11 +50,6 @@ class TestRepeatedGame:
 
     def test_best_deviation(self):
         ok_(best_dev_i(self.rpd, 0, 0) == 1)
-
-    def test_unit_circle(self):
-        H = RGUtil.unitcircle(4)
-        points = np.array([[1, 0], [0, 1], [-1, 0], [0, -1]])
-        ok_(max(abs(H - points).flatten()) < 1e-12)
 
     def test_subgradient_hyperplane_init(self):
         ok_(max(abs(self.C - np.ones(4))) < 1e-12)
