+++ quantecon/game_theory/utilities.py
@@ -51,63 +51,3 @@ class NashResult(dict):
 
     def __dir__(self):
         return self.keys()
-
-
-class RGUtil:
-    def frange(start, stop, step=1.):
-        """
-        Return evenly spaced floats within [start, stop].
-
-        Parameters
-        ----------
-        start : scalar(float), optional(default=0.)
-            Start of the interval.
-
-        stop : scalar(float)
-            End of the interval
-
-        step : scalar(float), optional(default=1.)
-            Spacing between values.
-
-        Returns
-        -------
-        Generator object
-
-        """
-        i = 0.0
-        x = float(start)  # Prevent yielding integers.
-        x0 = x
-        epsilon = step / 2.0
-        yield x  # yield always first value
-        while x + epsilon < stop:
-            i += 1.0
-            x = x0 + i * step
-            yield x
-
-    def unitcircle(npts):
-        """
-        Places `npts` equally spaced points along the 2 dimensional circle and 
-        returns the points with x coordinates in first column and y coordinates
-         in second column
-
-        Parameters
-        ----------
-        npts : scalar(float)
-
-        Returns
-        -------
-        ndarray(float, dim=2)
-
-        """
-        import math
-        import numpy as np
-
-        incr = 2 * math.pi / npts
-        degrees = [x for x in RGUtil.frange(0, 2*math.pi, incr)]
-        
-        pts = np.empty((npts, 2))
-        for i in range(npts):
-            x = degrees[i]
-            pts[i, 0] = math.cos(x)
-            pts[i, 1] = math.sin(x)
-        return pts
